Added code for Content tab "Manually Installed"
