+++ Source/Menu/MainForm.cs
@@ -578,7 +578,7 @@ namespace ORTS
         
         void buttonDownloadContent_Click(object sender, EventArgs e)
         {
-            using (var form = new DownloadContentForm(Settings))
+            using (var form = new ContentForm(Settings))
             {
                 form.ShowDialog(this);
             }
@@ -757,7 +757,7 @@ namespace ORTS
 
                 if (!initialized && Folders.Count == 0)
                 {
-                    using (var form = new DownloadContentForm(Settings))
+                    using (var form = new ContentForm(Settings))
                     {
                         switch (form.ShowDialog(this))
                         {
